views - added for productFiles

+++ src/main/webapp/WEB-INF/views/productFiles/all.jsp
@@ -3,40 +3,72 @@
 <html>
 <head>
     <title>Files</title>
-    <link rel="stylesheet" href="/css/styles.css">
+    <%@ include file="../dependecies.jsp" %>
 </head>
 <body>
+<%@include file="../loggedHeader.jsp" %>
 
-<button onclick="location.href='http://localhost:8080/'" type="button">HOME</button>
+<div class="w3-display-container w3-theme-l5" style="height:100%">
 
-<h3>Files:</h3>
+    <div class="w3-container w3-theme-l4" style="padding-top: 52px">
+        <h2>Product files list:</h2>
+    </div>
 
-<table border="1" style="text-align: center">
+    <div class="w3-container w3-margin">
 
-    <thead>
-    <th>LP</th>
-    <th>fileName</th>
-    <th>fileDescription</th>
-    <th colspan="3">Actions</th>
-    </thead>
+        <c:if test="${not empty param.error}">
+            <div class="w3-panel w3-pale-red w3-border w3-display-container">
+                        <span onclick="this.parentElement.style.display='none'"
+                              class="w3-button w3-large w3-display-topright">×</span>
+                <h4 class="w3-center">Item cannot be deleted - it has some records in database</h4>
+            </div>
+        </c:if>
 
-    <tbody>
-    <c:forEach items="${files}" var="file" varStatus="stat">
-        <tr>
-            <td>${stat.index+1}</td>
-            <td>${file.fileName}</td>
-            <td>${file.fileDescription}</td>
+        <c:if test="${not empty param.deleted}">
+            <div class="w3-panel w3-pale-green w3-border w3-display-container">
+                        <span onclick="this.parentElement.style.display='none'"
+                              class="w3-button w3-large w3-display-topright">×</span>
+                <h4 class="w3-center">Item has been deleted</h4>
+            </div>
+        </c:if>
 
-            <td style="width: 50px"><a href="/files/download/${file.id}" >Download</a></td>
-            <td style="width: 50px"><a href="/files/view/${file.id}" >View</a></td>
-            <td style="width: 50px"><a href="/files/delete/${file.id}" >Delete</a></td>
-        </tr>
-    </c:forEach>
-    </tbody>
+        <a href="http://localhost:8080/files/add/${productId}" class="w3-btn w3-medium w3-block w3-theme w3-large"
+           style="width:30%">Add a new item</a>
+    </div>
 
-</table>
+    <div class="w3-container w3-margin">
+        <table class="w3-table w3-striped w3-border w3-bordered w3-hoverable w3-card-4" style="width:30%">
 
-<h4><a href="http://localhost:8080/files/add/${productId}">Add a new file to this product</a></h4>
+            <thead>
+            <tr class="w3-light-grey">
+                <th>LP</th>
+                <th>File name</th>
+                <th>Description</th>
+                <th colspan="3">Actions</th>
+            </tr>
+            </thead>
+
+            <tbody>
+            <c:forEach items="${files}" var="file" varStatus="stat">
+                <tr>
+                    <td>${stat.index+1}</td>
+                    <td>${file.fileName}</td>
+                    <td>${file.fileDescription}</td>
+
+                    <td style="width: 50px"><a href="/files/download/${file.id}" >Download</a></td>
+                    <td style="width: 50px"><a href="/files/view/${file.id}" >View</a></td>
+                    <td style="width: 50px"><a href="/files/delete/${file.id}" onclick="return confirm('Are you sure you want to delete this item?');">Delete</a></td>
+                </tr>
+            </c:forEach>
+            </tbody>
+
+        </table>
+    </div>
+</div>
+
+<div style="position: relative; margin-top:-30.3px">
+    <%@include file="../footer.jsp" %>
+</div>
 
 </body>
 </html>
